requirements.txt
praw
wget
git+https://git@github.com/ping/instagram_private_api.git@1.6.0
Pillow
